src/nagiosql/requirements_test.txt
coverage==3.5.2
django-jenkins==0.13.0
PyYAML==3.10

